// src/app/models/treatment.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TreatmentDuration } from './treatment-duration';


export interface Treatment { 
    createdDate?: string;
    description?: string;
    durations?: Array<TreatmentDuration>;
    id?: number;
    lastModifiedBy?: string;
    lastModifiedDate?: string;
    name?: number;
}